Fix line-numbering bug
Python at its core is zero-indexed. I think that's why a call to `enumerate` inside a for-loop will start at index 0.

BUT, pygments assumes the first line of code input is 1.

SO, The highlighted ('changed') lines in the timeline render weren't matching up with what actually changed in that commit; they were one off.

In summary: off-by-one errors are frustrating.