Properly shrink snapshots on zoom
Before, the syntax-highlighted snapshots had this like, extra wide margin on the right-hand side because (unbeknownst to me) I was using a crude `<table>` selector (to select the snapshots table), but it also applied to the pygments syntax-highlighted output (which is itself a table). So it's tables within tables really.